26-02-19

+++ src/main/java/com/example/smartparkingsystem/DB/validation.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS `validation`
 (
-    `no`           INT PRIMARY KEY AUTO_INCREMENT COMMENT '인덱스',
+    `no`           BIGINT PRIMARY KEY AUTO_INCREMENT COMMENT '인덱스',
     `admin_id`     VARCHAR(20) NOT NULL COMMENT 'FK 관리자 아이디',
     `otp_code`     CHAR(6)     NOT NULL COMMENT 'OTP',
     `admin_email`  VARCHAR(50) NOT NULL COMMENT '관리자 이메일 (OTP보낸 이메일)',
